Support for Status-type interrupts
Extend DMA DV to support Status-type interrupts and
not just Event-type interrupts.

Invalidate all pending interrupt predictions when a
DMA transfer is aborted because there is a window between
the final TL-UL write completing and signaling completion
of the DMA transfer during which an abort may occur,
preventing the interrupt change.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>