Changes to CFG have immediate effect
Writes to CFG immediately reset the shared phase counter
logic, and this is important because otherwise it would
be possible to put the logic into a state that it is
unresponsive for a very long time.

Ensure that the phase counter responds to reconfiguration
at any point too, and not just at the end of each beat.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>