[dv,darjeeling] Support sw_symbol_backdoor_write
Implement sw_symbol_backdoor_write for scrambled and non-
scrambled RAM memories; this was not previously possible but
the SW test image lives in a RAM on the CTN network in
Darjeeling.

Presently that RAM is not scrambled, but it should be changed
to a regular scrambled RAM instance in the future.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>