Initial work on I3C IP block
The test bench and register interface are together capable of
reading from/writing to the internal message buffer.

Currently the logic is capable of transmitting, synchronizing
and receiving words read from/to the message buffer.

Transmission and reception occur in both directions, using
SDR signalling but this is not legitimate SDR traffic at
present; there is no arbitration, no T bit and the 32-bit
word length is not valid SDR traffic. The logic is presently
geared more towards HDR-DDR but this requires HDR Exit
signalling prior to the generation of SDR STOP signalling.

Main IP clock is 50MHz and the AON clock is targeting 100kHz.